server_api/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.5
Django==4.2.7
django-cors-headers==4.3.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
filelock==3.13.1
fsspec==2023.10.0
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
pycparser==2.21
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
sqlparse==0.4.4
sympy==1.12
tokenizers==0.14.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.66.1
transformers==4.35.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
kaggle==1.6.14
scikit-learn==1.5.0
pandas==2.2.2
seaborn==0.13.2